refactor(automation-hub): reorganize components and improve database security
- Replace check constraints with row-level security policies for multi-tenant isolation
- Consolidate automation hub pages into a single tab-based component structure
- Fix tenant handling in workflow templates seed file
- Improve database connection handling by removing manual knex.destroy() calls
- Add WorkflowRegistrationModel to shared code for better abstraction
- Fix filename for event catalog migration
- Optimize database queries with JOINs to avoid N+1 problems
- Improve code organization with proper component separation
