+++ server/src/lib/actions/workflow-runtime-actions.ts
@@ -6,22 +6,12 @@
 import { createTenantKnex } from 'server/src/lib/db';
 import { Knex } from 'knex';
 import logger from '@shared/core/logger.js';
-
-/**
- * Interface for a workflow registration
- */
-interface WorkflowRegistration {
-  registration_id: string;
-  name: string;
-  version: string;
-  definition: any;
-  parameters?: any;
-}
+import WorkflowRegistrationModel, { WorkflowRegistration } from '@shared/workflow/persistence/workflowRegistrationModel.js';
 
 /**
  * Get a workflow registration by name and optional version
  * If version is not provided, returns the current version
- * 
+ *
  * @param name The workflow name
  * @param version Optional version string
  * @returns The workflow registration or null if not found
@@ -29,49 +19,18 @@ interface WorkflowRegistration {
 export async function getWorkflowRegistration(name: string, version?: string): Promise<WorkflowRegistration | null> {
   const { knex, tenant } = await createTenantKnex();
   
+  if (!tenant) {
+    throw new Error('Tenant is required');
+  }
+  
   try {
-    // Get the workflow registration
-    const registration = await knex('workflow_registrations')
-      .select('registration_id', 'name', 'version')
-      .where('name', name)
-      .where('tenant_id', tenant)
-      .first();
-      
-    if (!registration) {
-      return null;
-    }
-    
-    // Get the specific version or current version
-    let versionQuery = knex('workflow_registration_versions')
-      .where('registration_id', registration.registration_id);
-      
-    if (version) {
-      versionQuery = versionQuery.where('tenant_id', tenant);
-      versionQuery = versionQuery.where('version', version);
-    } else {
-      versionQuery = versionQuery.where('tenant_id', tenant);
-      versionQuery = versionQuery.where('is_current', true);
-    }
-    
-    const versionRecord = await versionQuery.first();
-    
-    if (!versionRecord) {
-      return null;
-    }
-    
-    return {
-      ...registration,
-      definition: versionRecord.definition,
-      parameters: versionRecord.parameters
-    };
+    // Use the model to get the workflow registration
+    return await WorkflowRegistrationModel.getByName(knex, tenant, name, version);
   } catch (error) {
     logger.error(`Error getting workflow registration for ${name}:`, error);
     throw error;
   } finally {
-    // Release the knex connection
-    if (knex) {
-      await knex.destroy();
-    }
+    // Connection will be released automatically
   }
 }
 
@@ -83,45 +42,18 @@ export async function getWorkflowRegistration(name: string, version?: string): P
 export async function getAllWorkflowRegistrations(): Promise<WorkflowRegistration[]> {
   const { knex, tenant } = await createTenantKnex();
   
+  if (!tenant) {
+    throw new Error('Tenant is required');
+  }
+  
   try {
-    // Get all active workflow registrations
-    const registrations = await knex('workflow_registrations')
-      .where('tenant_id', tenant)
-      .select('registration_id', 'name', 'version')
-      .where('status', 'active');
-    
-    if (!registrations.length) {
-      return [];
-    }
-    
-    // Get the current version for each registration
-    const results: WorkflowRegistration[] = [];
-    
-    for (const registration of registrations) {
-      const versionRecord = await knex('workflow_registration_versions')
-        .where('tenant_id', tenant)
-        .where('registration_id', registration.registration_id)
-        .where('is_current', true)
-        .first();
-      
-      if (versionRecord) {
-        results.push({
-          ...registration,
-          definition: versionRecord.definition,
-          parameters: versionRecord.parameters
-        });
-      }
-    }
-    
-    return results;
+    // Use the model to get all workflow registrations
+    return await WorkflowRegistrationModel.getAll(knex, tenant);
   } catch (error) {
     logger.error('Error getting all workflow registrations:', error);
     throw error;
   } finally {
-    // Release the knex connection
-    if (knex) {
-      await knex.destroy();
-    }
+    // Connection will be released automatically
   }
 }
 
@@ -137,57 +69,24 @@ export async function createRegistrationFromTemplate(params: {
   description?: string;
   parameters?: any;
 }): Promise<{ registrationId: string }> {
-  const { templateId, name, description, parameters } = params;
   const { knex, tenant } = await createTenantKnex();
   
+  if (!tenant) {
+    throw new Error('Tenant is required');
+  }
+  
   try {
-    return await knex.transaction(async (trx: Knex.Transaction) => {
-      // Get the template
-      const template = await trx('workflow_templates')
-        .where('tenant_id', tenant)
-        .where('template_id', templateId)
-        .first();
-      
-      if (!template) {
-        throw new Error(`Template with ID ${templateId} not found`);
-      }
-      
-      // Create the registration
-      const [registration] = await trx('workflow_registrations')
-        .insert({
-          tenant_id: tenant,
-          name,
-          description: description || template.description,
-          category: template.category,
-          tags: template.tags,
-          version: '1.0.0',
-          status: 'active',
-          source_template_id: templateId,
-          definition: template.definition,
-          parameters: parameters || template.default_parameters || {}
-        })
-        .returning('registration_id');
-      
-      // Create the initial version
-      await trx('workflow_registration_versions')
-        .insert({
-          tenant_id: tenant,
-          registration_id: registration.registration_id,
-          version: '1.0.0',
-          is_current: true,
-          definition: template.definition,
-          parameters: parameters || template.default_parameters || {}
-        });
-      
-      return { registrationId: registration.registration_id };
+    // Use the model to create a registration from a template
+    return await WorkflowRegistrationModel.createFromTemplate(knex, tenant, {
+      templateId: params.templateId,
+      name: params.name,
+      description: params.description,
+      parameters: params.parameters
     });
   } catch (error) {
-    logger.error(`Error creating workflow registration from template ${templateId}:`, error);
+    logger.error(`Error creating workflow registration from template ${params.templateId}:`, error);
     throw error;
   } finally {
-    // Release the knex connection
-    if (knex) {
-      await knex.destroy();
-    }
+    // Connection will be released automatically
   }
 }
